finish addressing feedback comments and fix unit tests
