+++ pyrit/score/true_false/video_true_false_scorer.py
@@ -5,7 +5,6 @@ from typing import Optional
 
 from pyrit.identifiers import ScorerIdentifier
 from pyrit.models import MessagePiece, Score
-from pyrit.score.audio_transcript_scorer import _BaseAudioTranscriptScorer
 from pyrit.score.scorer_prompt_validator import ScorerPromptValidator
 from pyrit.score.true_false.true_false_score_aggregator import (
     TrueFalseAggregatorFunc,
@@ -105,9 +104,11 @@ class VideoTrueFalseScorer(TrueFalseScorer, _BaseVideoScorer):
 
         # Score audio if audio_scorer is provided
         if self.audio_scorer:
-            audio_score = await self._score_video_audio_async(message_piece=message_piece, objective=objective)
-            if audio_score:
-                all_scores.append(audio_score)
+            audio_scores = await self._score_video_audio_async(
+                message_piece=message_piece, audio_scorer=self.audio_scorer
+            )
+            if audio_scores:
+                all_scores.extend(audio_scores)
                 audio_scored = True
 
         # Use the TrueFalseAggregatorFunc to combine all scores (frames + audio)
@@ -117,10 +118,10 @@ class VideoTrueFalseScorer(TrueFalseScorer, _BaseVideoScorer):
         piece_id = message_piece.id if message_piece.id is not None else message_piece.original_prompt_id
 
         # Build rationale
-        rationale_parts = [f"Video scored by analyzing {len(frame_scores)} frames"]
+        rationale = f"Video scored by analyzing {len(frame_scores)} frames"
         if audio_scored:
-            rationale_parts.append("and audio transcript")
-        rationale_parts.append(f".\n{result.rationale}")
+            rationale += " and audio transcript"
+        rationale += f".\n{result.rationale}"
 
         # Create the aggregated score using the aggregator result
         aggregate_score = Score(
@@ -129,75 +130,10 @@ class VideoTrueFalseScorer(TrueFalseScorer, _BaseVideoScorer):
             score_type="true_false",
             score_category=result.category,
             score_metadata=result.metadata,
-            score_rationale="".join(rationale_parts),
+            score_rationale=rationale,
             scorer_class_identifier=self.get_identifier(),
             message_piece_id=piece_id,
             objective=objective,
         )
 
         return [aggregate_score]
-
-    async def _score_video_audio_async(
-        self, *, message_piece: MessagePiece, objective: Optional[str] = None
-    ) -> Optional[Score]:
-        """
-        Extract and score audio from the video.
-
-        Args:
-            message_piece: The message piece containing the video.
-            objective: Optional objective description for scoring.
-
-        Returns:
-            Score for the audio content, or None if audio extraction/scoring fails.
-        """
-        import os
-        import uuid
-
-        from pyrit.memory import CentralMemory
-
-        video_path = message_piece.converted_value
-
-        # Use _BaseAudioTranscriptScorer's static method to extract audio
-        audio_path = _BaseAudioTranscriptScorer.extract_audio_from_video(video_path)
-        if not audio_path:
-            return None
-
-        try:
-            # Create a message piece for the audio
-            original_prompt_id = message_piece.original_prompt_id
-            if isinstance(original_prompt_id, str):
-                original_prompt_id = uuid.UUID(original_prompt_id)
-
-            audio_piece = MessagePiece(
-                original_value=audio_path,
-                role=message_piece.get_role_for_storage(),
-                original_prompt_id=original_prompt_id,
-                converted_value=audio_path,
-                converted_value_data_type="audio_path",
-            )
-
-            audio_message = audio_piece.to_message()
-
-            # Add to memory
-            memory = CentralMemory.get_memory_instance()
-            memory.add_message_to_memory(request=audio_message)
-
-            # Score the audio using the audio_scorer
-            # We pass objective=None because when used within a video scorer,
-            # the audio should be evaluated independently using only its true_description,
-            # not the overall video objective (which describes visual elements)
-            if self.audio_scorer is None:
-                return None
-            audio_scores = await self.audio_scorer.score_prompts_batch_async(
-                messages=[audio_message],
-                objectives=None,  # Audio uses only its own true_description, not video objective
-                batch_size=1,
-            )
-
-            return audio_scores[0] if audio_scores else None
-
-        finally:
-            # Clean up temporary audio file
-            if os.path.exists(audio_path):
-                os.unlink(audio_path)
-            pass
